feat(registry): extract capability assembly into a tauri-free crate
